+++ acrg/obs/process_gcwerks_dagage2.sh
@@ -2,21 +2,12 @@
 # Run this before process_gcwerks, to get the most up-to-date data
 
 # AGAGE GC
-rsync -avh dagage2.chm.bris.ac.uk:/agage/summary/data/* /work/chxmr/shared/obs_raw/AGAGE_GCWerks/data/
-rsync -avh dagage2.chm.bris.ac.uk:/agage/summary/data-gcms/* /work/chxmr/shared/obs_raw/AGAGE_GCWerks/data-gcms/
-rsync -avh dagage2.chm.bris.ac.uk:/agage/summary/air-history/baseline_flags/* /work/chxmr/shared/obs_raw/AGAGE_GCWerks/baseline_flags/
+rsync -avh dagage2.chm.bris.ac.uk:/agage/summary/data/* /user/work/chxmr/shared/obs_raw/AGAGE_GCWerks/data/
+rsync -avh dagage2.chm.bris.ac.uk:/agage/summary/data-gcms/* /user/work/chxmr/shared/obs_raw/AGAGE_GCWerks/data-gcms/
+rsync -avh dagage2.chm.bris.ac.uk:/agage/summary/air-history/baseline_flags/* /user/work/chxmr/shared/obs_raw/AGAGE_GCWerks/baseline_flags/
 
-# DECC GC
-sites="barbados bristol bilsdale heathfield macehead ridgehill tacolneston"
-for site in $sites
-do
-    #  first, copy Mace Head MD (which I think is the only one with site followed by a period)
-    rsync -avh dagage2.chm.bris.ac.uk:/agage/summary/data-net/${site}.* /work/chxmr/shared/obs_raw/AGAGE_GCWerks/data/
-    #  then, copy over files labeled -md (CHECK: Are these also labeled as -md in the data folder? Think so)
-    rsync -avh dagage2.chm.bris.ac.uk:/agage/summary/data-net/${site}-md* /work/chxmr/shared/obs_raw/AGAGE_GCWerks/data/
-    #  Copy over medusa files
-    rsync -avh dagage2.chm.bris.ac.uk:/agage/summary/data-net/${site}-medusa* /work/chxmr/shared/obs_raw/AGAGE_GCWerks/data-gcms/
-done
+# DECC GCMD
+rsync -avh dagage2.chm.bris.ac.uk:/agage/summary/gccompare-decc/data/md/* /user/work/chxmr/shared/obs_raw/AGAGE_GCWerks/data/
 
 # DECC/GAUGE/AGAGE CRDS
 folders=`ssh dagage2.chm.bris.ac.uk "ls /agage/"`
@@ -27,14 +18,17 @@ do
     for site in $sites
     do
         if [[ $folder == *"$site-picarro"* ]]; then
-            rsync -avh dagage2.chm.bris.ac.uk:/agage/${folder}/results-gcwerks/*.dat /work/chxmr/shared/obs_raw/AGAGE_GCWerks/data-picarro/$site/
+            rsync -avh dagage2.chm.bris.ac.uk:/agage/${folder}/results-gcwerks/*.dat /user/work/chxmr/shared/obs_raw/AGAGE_GCWerks/data-picarro/$site/
         fi
     done
 done
 
+# Get HFD 5310
+rsync -avh dagage2.chm.bris.ac.uk:/agage/heathfield-5310/results-gcwerks/*.dat /user/work/chxmr/shared/obs_raw/AGAGE_GCWerks/data-picarro/heathfield/
+
 # Get TAC LGR
-rsync -avh dagage2.chm.bris.ac.uk:/agage/tacolneston-lgr/results-gcwerks/*.dat /work/chxmr/shared/obs_raw/AGAGE_GCWerks/data-picarro/tacolneston/
+rsync -avh dagage2.chm.bris.ac.uk:/agage/tacolneston-lgr/results-gcwerks/*.dat /user/work/chxmr/shared/obs_raw/AGAGE_GCWerks/data-picarro/tacolneston/
 
 # Copy ICOS data
-rsync -avh dagage2.chm.bris.ac.uk:/agage/macehead-picarro/results-icos/* /work/chxmr/shared/obs_raw/AGAGE_GCWerks/data-icos/macehead/
-rsync -avh dagage2.chm.bris.ac.uk:/agage/angus-picarro/results-icos/* /work/chxmr/shared/obs_raw/AGAGE_GCWerks/data-icos/angus/
+rsync -avh dagage2.chm.bris.ac.uk:/agage/macehead-picarro/results-icos/* /user/work/chxmr/shared/obs_raw/AGAGE_GCWerks/data-icos/macehead/
+rsync -avh dagage2.chm.bris.ac.uk:/agage/angus-picarro/results-icos/* /user/work/chxmr/shared/obs_raw/AGAGE_GCWerks/data-icos/angus/